Add files via upload
Updating the code to disable the fixed apd alternan option when the user is analyzing a calcium image. It greys out the option (option 3) and it provides a 'tooltip' popup upon hover, so the user knows why they cannot edit the features. It also has an editable max amp feature for the s1-s2 analysis, in case the s2 is too short, the user can reduce or increase threshold.